Moved static items into files.
